pyproject.toml
[project]
name = "mcp-visualizer"
version = "0.1.0"
description = "An educational tool that visualizes Model Context Protocol (MCP) communication between an LLM-powered chat interface and MCP servers, designed to teach users how MCP orchestrates tool calling through transparent, real-time message flow visualization."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.2.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
